+++ app/services/local-data.ts
@@ -5,51 +5,22 @@ import ManagerService from './manager';
 import { clean, valid, gt } from 'semver';
 import NewsFeedService from './news-feed';
 
-const DATA_COOKIE_NAME = 'doleo-data';
-
 export interface LocalData {
   latestEncounteredVersion: string;
 }
 
 export default class LocalDataService extends Service {
-  @service declare cookies: any;
   @service declare manager: ManagerService;
   @service declare newsFeed: NewsFeedService;
 
-  data: LocalData = {
-    latestEncounteredVersion: 'unknown',
-  };
-
-  constructor() {
-    super(...arguments);
-    this.read();
-  }
-
-  write(data: LocalData) {
-    this.cookies.write(DATA_COOKIE_NAME, JSON.stringify(data), {
-      sameSite: 'strict',
-    });
-  }
-
-  read() {
-    const cookie = this.cookies.read(DATA_COOKIE_NAME);
-    try {
-      if (cookie) {
-        this.data = JSON.parse(cookie);
-      } else {
-        this.write(this.data);
-      }
-    } catch (error) {
-      this.write(this.data);
-    }
-  }
-
   getUnencounteredVersion(): string | undefined {
+    const encounteredVersion =
+      localStorage.getItem('doleo-encounteredVersion') || '';
     if (
-      !valid(this.data.latestEncounteredVersion) ||
+      !valid(encounteredVersion) ||
       gt(
         clean(ENV.APP['version'] as string) as string,
-        clean(this.data.latestEncounteredVersion) as string
+        clean(encounteredVersion) as string
       )
     ) {
       return clean(ENV.APP['version'] as string) as string;
@@ -57,11 +28,11 @@ export default class LocalDataService extends Service {
     return undefined;
   }
 
-  @action rememberEncounteredVersion() {
-    this.write({
-      ...this.data,
-      latestEncounteredVersion: (ENV.APP['version'] as string) || '',
-    });
+  @action setEncounteredVersion() {
+    localStorage.setItem(
+      'doleo-encounteredVersion',
+      (ENV.APP['version'] as string) || ''
+    );
     this.newsFeed.newVersion = undefined;
     this.newsFeed.update();
   }